+++ src/.internal/core/render/MonotoneXTension.ts
@@ -44,7 +44,7 @@ export class MonotoneXTension {
   }
 
   public point(x: number, y: number) {
-    var t1 = NaN;
+    let t1 = NaN;
 
     x = +x, y = +y;
     if (x === this._x1 && y === this._y1) return; // Ignore coincident points.
@@ -96,7 +96,7 @@ function slope3(that: MonotoneXTension, x2: number, y2: number) {
  * @ignore
  */
 function slope2(that: MonotoneXTension, t: number) {
-  var h = that._x1 - that._x0;
+  let h = that._x1 - that._x0;
   return h ? (3 * (that._y1 - that._y0) / h - t) / 2 : t;
 }
 
@@ -104,7 +104,7 @@ function slope2(that: MonotoneXTension, t: number) {
  * @ignore
  */
 function point(that: MonotoneXTension, t0: number, t1: number) {
-  var x0 = that._x0,
+  let x0 = that._x0,
     y0 = that._y0,
     x1 = that._x1,
     y1 = that._y1,